Fix invalid SQL exception on library path change
In case the music library path was changed to sub directory of the
previous directory, but all the previously scanned files resided in the
new directory, the Scanner tried to delete an empty set of tracks. This
lead to executing invalid SQL statement "WHERE file_id IN ()" and an
exception was thrown. Hence, the HTTP POST '/api/settings/user/path'
returned 500 and the UI state did not update correctly.

+++ utility/scanner.php
@@ -489,7 +489,9 @@ class Scanner extends PublicEmitter {
 		$indexedFiles = $this->getScannedFiles($userId);
 		$validFiles = Util::extractIds($this->getMusicFiles($userId, $userHome));
 		$filesToRemove = Util::arrayDiff($indexedFiles, $validFiles);
-		$this->deleteAudio($filesToRemove, [$userId]);
+		if (\count($filesToRemove)) {
+			$this->deleteAudio($filesToRemove, [$userId]);
+		}
 	}
 
 	/**
